requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
twilio==8.10.0
openai
pymongo==4.6.0
boto3==1.34.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
langchain==0.0.340
langchain-openai==0.0.2
langchain-community==0.0.10
faiss-cpu
sentence-transformers==2.2.2
boto3